Add generic `Name`s and `Value`s in the types of `Send`, `Receive`, etc.
Notice this modifies slightly the macro for the #[session] attribute for
enums: one should now provide the types of names and values.

E.g. #[session((), ())] for an enum that does not use the map.
See `examples/3buyers.rs` for an example. I expect to extend this
approach (to pass the types as macro parameters) in the future.